requirements.txt
discord.py==1.7.3
python-dotenv==0.19.1
requests==2.26.0
openai==0.10.5